keeper.c: fix unsigned int cast of pointer

+++ src/keeper.c
@@ -625,7 +625,7 @@ struct s_Keeper* keeper_acquire( void const* ptr)
 		*
 		* Pointers are often aligned by 8 or so - ignore the low order bits
 		*/
-		unsigned int i = ((unsigned int)(ptr) >> 3) % GNbKeepers;
+		unsigned long i = ((unsigned long)(ptr) >> 3) % GNbKeepers;
 		struct s_Keeper *K= &GKeepers[i];
 
 		MUTEX_LOCK( &K->lock_);
